requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
gspread>=6.0.0
google-auth>=2.25.0
flask>=3.0.0
